test: add dedicated tests for partitions.py
Extracted tests from monolithic `test_DiscreteMath.py` into a dedicated file
`test_partitions.py` and achieved 100% test coverage for
`Math/Discrete_Math/Number_Theory/partitions.py`.
